+++ modules/vue-cli-plugin-tbox/generator/templates/default/src/index.ts
@@ -1,25 +1,15 @@
-import { Application } from '@t-box/server';
 import http from 'http';
 import boot from './server';
 
-const server = http.createServer();
-let currentApp: Application;
+let server = http.createServer();
 
-let port = parseInt(process.env.PORT as string, 10);
-let hostname = process.env.HOSTNAME;
+const port = parseInt(process.env.PORT as string, 10);
+const hostname = process.env.HOSTNAME;
 
-server.listen(port, hostname);
-
-boot()
-  .then(app => {
+boot(server)
+  .then(() => {
     console.log(`Binding app to ${hostname}:${port}`);
-    app.bootstrap(server)
-      .then(() => {
-        currentApp = app;
-      }).catch(err => {
-        console.log(err);
-        process.exit(1);
-      });
+    server.listen(port, hostname);
   }).catch(err => {
     console.log(err);
     process.exit(1);
@@ -27,16 +17,12 @@ boot()
 
 if (module.hot) {
   module.hot.accept('./server', () => {
-    boot()
-      .then(app => {
-        currentApp.unbind(server);
-        app.bootstrap(server)
-          .then(() => {
-            currentApp = app;
-          }).catch(err => {
-            console.log(err);
-            process.exit(1);
-          });
+    server.close();
+    server = http.createServer();
+    boot(server)
+      .then(() => {
+        console.log(`Binding app to ${hostname}:${port}`);
+        server.listen(port, hostname);
       }).catch(err => {
         console.log(err);
         process.exit(1);