+++ modules/vue-cli-plugin-tbox/generator/templates/default/src/server.ts
@@ -1,7 +1,8 @@
+import { Server } from 'http';
 import { Application } from '@t-box/server';
 
-export default async function boot() {
+export default async function boot(server: Server) {
   const app = new Application({ cwd: __dirname, staticFolders: ['public'], assetDir: 'public', compress: true });
   await app.startup(require.context('./modules', true, /startup\.ts$/));
-  return app;
+  await app.bootstrap(server);
 }